rename: global -> globals 디렉토리명 수정 (#78)
