+++ src/globals/styles/global.css
