+++ src/globals/types/svg.d.ts
